Split apart input components
